+++ adapters/IronSource/IronSourceAdapter/GADMAdapterIronSourceConstants.h
@@ -14,8 +14,9 @@
 
 // IronSource internal reporting const.
 static NSString *const kGADMAdapterIronSourceMediationName = @"AdMob";
-static NSString *const kGADMAdapterIronSourceAdapterVersion = @"310";
+static NSString *const kGADMAdapterIronSourceAdapterVersion = @"6.8.4.1";
 static NSString *const kGADMIronSourceDefaultInstanceId = @"0";
+static NSString *const kGADMAdapterIronSourceInternalVersion = @"310";
 
 // IronSource parameters keys.
 static NSString *const kGADMAdapterIronSourceAppKey = @"appKey";