+++ adapters/Oath/OathAdapter/GADMediationAdapterVerizon.h
@@ -1,0 +1,11 @@
+//
+//  GADMediationAdapterVerizon.h
+//
+// @copyright Copyright (c) 2018 Verizon. All rights reserved.
+//
+
+#import "GADMVASAdapterBaseClass.h"
+
+@interface GADMediationAdapterVerizon : GADMVASAdapterBaseClass
+
+@end