+++ adapters/Tapjoy/TapjoyAdapter/GADMAdapterTapjoy.m
@@ -19,7 +19,7 @@
 #import "GADMTapjoyExtras.h"
 #import "GADMediationAdapterTapjoy.h"
 
-@interface GADMAdapterTapjoy () <TJPlacementVideoDelegate, TJPlacementDelegate> {
+@interface GADMAdapterTapjoy () <TJPlacementDelegate, TJPlacementVideoDelegate> {
   // Connector from Google Mobile Ads SDK to receive ad configurations.
   __weak id<GADMAdNetworkConnector> _interstitialConnector;
 
@@ -146,6 +146,26 @@
   id<GADMAdNetworkConnector> strongConnector = _interstitialConnector;
   [strongConnector adapterWillDismissInterstitial:self];
   [strongConnector adapterDidDismissInterstitial:self];
+}
+
+- (void)didClick:(TJPlacement*)placement {
+  id<GADMAdNetworkConnector> strongConnector = _interstitialConnector;
+  [strongConnector adapterDidGetAdClick:self];
+  [strongConnector adapterWillLeaveApplication:self];
+}
+
+#pragma mark Tapjoy Video
+
+- (void)videoDidStart:(TJPlacement *)placement {
+  // Do nothing
+}
+
+- (void)videoDidComplete:(TJPlacement *)placement {
+  // Do nothing
+}
+
+- (void)videoDidFail:(TJPlacement *)placement error:(NSString *)errorMsg {
+  // Do nothing
 }
 
 @end