+++ adapters/DuAd/CHANGELOG.md
@@ -1,5 +1,9 @@
 # DuAd Audience Network Mediation Adapter for Google Mobile Ads SDK for iOS
 
+## Next Version
+- Removed support for Native App Install ad requests. Apps must use the Unified Native Ads API to request native ads.
+- Now requires Google Mobile Ads SDK version 7.46.0 or higher.
+
 ## Version 1.1.2.1
 - Added code to initialize DuAd SDK before requesting ads.